# artifacts/api-server/src/routes/stocks.ts
import { Router } from "express";
import YahooFinanceClass from "yahoo-finance2";
const yahooFinance = new YahooFinanceClass();
import {
  CompareStocksQueryParams,
  SearchStocksQueryParams,
} from "@workspace/api-zod";

const router = Router();

interface CacheEntry<T> {
  value: T;
  expiresAt: number;
}

class TTLCache<T> {
  private store = new Map<string, CacheEntry<T>>();
  constructor(private ttlMs: number) {}
  get(key: string): T | undefined {
    const entry = this.store.get(key);
    if (!entry) return undefined;
    if (Date.now() > entry.expiresAt) { this.store.delete(key); return undefined; }
    return entry.value;
  }
  set(key: string, value: T) {
    this.store.set(key, { value, expiresAt: Date.now() + this.ttlMs });
  }
}

const searchCache = new TTLCache<object[]>(5 * 60 * 1000);
const compareCache = new TTLCache<object>(10 * 60 * 1000);

function safeNum(val: unknown): number | null {
  if (val === null || val === undefined || typeof val === "object") return null;
  const n = Number(val);
  return isFinite(n) ? n : null;
}

function classifyStockType(
  peRatio: number | null,
  pegRatio: number | null,
  dividendYield: number | null,
  revenueGrowth: number | null
): string {
  if (dividendYield !== null && dividendYield > 0.03) return "Dividend";
  if (pegRatio !== null && pegRatio < 1.5 && (peRatio === null || peRatio < 20))
    return "Value";
  if (revenueGrowth !== null && revenueGrowth > 0.15) return "Growth";
  if (peRatio !== null && peRatio > 30) return "Growth";
  return "Blend";
}

function computeFairValue(
  eps: number | null,
  epsGrowth: number | null,
  peRatio: number | null
): number | null {
  if (eps === null || epsGrowth === null) return null;
  const growthPct = epsGrowth * 100;
  const fairPE = 2 * growthPct;
  if (fairPE <= 0) return null;
  return parseFloat((eps * fairPE).toFixed(2));
}

function scoreMetric(
  val1: number | null,
  val2: number | null,
  higherIsBetter: boolean
): [number, number] {
  if (val1 === null && val2 === null) return [0, 0];
  if (val1 === null) return [0, 1];
  if (val2 === null) return [1, 0];
  if (val1 === val2) return [0.5, 0.5];
  if (higherIsBetter) {
    return val1 > val2 ? [1, 0] : [0, 1];
  } else {
    return val1 < val2 ? [1, 0] : [0, 1];
  }
}

function fmtPct(val: number | null): string | null {
  if (val === null) return null;
  return `${(val * 100).toFixed(2)}%`;
}

function fmtNum(val: number | null, decimals = 2): string | null {
  if (val === null) return null;
  return val.toFixed(decimals);
}

function fmtLargeNum(val: number | null): string | null {
  if (val === null) return null;
  if (Math.abs(val) >= 1e12) return `$${(val / 1e12).toFixed(2)}T`;
  if (Math.abs(val) >= 1e9) return `$${(val / 1e9).toFixed(2)}B`;
  if (Math.abs(val) >= 1e6) return `$${(val / 1e6).toFixed(2)}M`;
  return `$${val.toFixed(2)}`;
}

function fmtPrice(val: number | null): string | null {
  if (val === null) return null;
  return `$${val.toFixed(2)}`;
}

function buildScorecard(
  s1: ReturnType<typeof buildMetrics>,
  s2: ReturnType<typeof buildMetrics>,
  ticker1: string,
  ticker2: string
) {
  type ScorecardItem = {
    metric: string;
    weight: number;
    ticker1Score: number;
    ticker2Score: number;
    ticker1Value: string | null;
    ticker2Value: string | null;
    winner: string | null;
    explanation: string;
  };

  const items: ScorecardItem[] = [];

  function addItem(
    metric: string,
    weight: number,
    val1: number | null,
    val2: number | null,
    higherIsBetter: boolean,
    fmtFn: (v: number | null) => string | null,
    explanation: string
  ) {
    const [sc1, sc2] = scoreMetric(val1, val2, higherIsBetter);
    let winner: string | null = null;
    if (sc1 > sc2) winner = ticker1;
    else if (sc2 > sc1) winner = ticker2;
    items.push({
      metric,
      weight,
      ticker1Score: sc1,
      ticker2Score: sc2,
      ticker1Value: fmtFn(val1),
      ticker2Value: fmtFn(val2),
      winner,
      explanation,
    });
  }

  // ── Tier 1: Valuation (highest weight) ──────────────────────────────────
  // PEG is the gold standard: combines P/E with growth, penalising expensive growth
  addItem("PEG Ratio", 3.0, s1.pegRatio, s2.pegRatio, false, fmtNum,
    "The single best valuation shorthand. PEG < 1 = undervalued growth; > 2 = expensive. Invented by Peter Lynch.");

  // Price-to-FCF: harder to manipulate than earnings; Buffett's preferred metric
  const p2fcf1 = s1.marketCap !== null && s1.freeCashFlow !== null && s1.freeCashFlow > 0
    ? s1.marketCap / s1.freeCashFlow : null;
  const p2fcf2 = s2.marketCap !== null && s2.freeCashFlow !== null && s2.freeCashFlow > 0
    ? s2.marketCap / s2.freeCashFlow : null;
  addItem("Price / Free Cash Flow", 3.0, p2fcf1, p2fcf2, false, fmtNum,
    "How many dollars you pay per $1 of free cash flow. FCF is harder to manipulate than net income — pros treat this as the real earnings multiple.");

  // Analyst consensus upside: aggregated professional opinion on fair value
  const upside1 = s1.analystTargetPrice !== null && s1.currentPrice !== null
    ? (s1.analystTargetPrice / s1.currentPrice) - 1 : null;
  const upside2 = s2.analystTargetPrice !== null && s2.currentPrice !== null
    ? (s2.analystTargetPrice / s2.currentPrice) - 1 : null;
  addItem("Analyst Upside to Target", 2.5, upside1, upside2, true, fmtPct,
    "Wall Street consensus target price vs. today's price. Positive = analysts see upside; negative = stock is above consensus fair value.");

  // Trailing P/E: familiar benchmark, but penalise vs PEG for high-growth stocks
  addItem("P/E Ratio", 2.0, s1.peRatio, s2.peRatio, false, fmtNum,
    "Price-to-earnings: how much investors pay per $1 of trailing profit. Useful for stable earners; less meaningful for high-growth companies (use PEG instead).");

  // ── Tier 2: Growth quality ───────────────────────────────────────────────
  addItem("Revenue Growth (YoY)", 2.5, s1.revenueGrowthYoY, s2.revenueGrowthYoY, true, fmtPct,
    "Year-over-year revenue growth — the top-line engine. Sustained revenue growth above 20% is exceptional; below 5% for a growth stock is a red flag.");

  addItem("EPS Growth", 2.5, s1.epsGrowth, s2.epsGrowth, true, fmtPct,
    "Earnings-per-share growth (trailing vs. forward). The compounding driver of long-term stock price. Rising EPS with rising margins = quality growth.");

  // ── Tier 3: Profitability ────────────────────────────────────────────────
  addItem("Net Profit Margin", 2.0, s1.netMargin, s2.netMargin, true, fmtPct,
    "How many cents of every revenue dollar become profit. Buffett looks for durable margins >15%. Expanding margins over time = pricing power.");

  addItem("Return on Equity (ROE)", 2.0, s1.returnOnEquity, s2.returnOnEquity, true, fmtPct,
    "Profit generated per dollar of shareholder equity. ROE > 15% consistently signals a durable competitive advantage. Buffett's key quality screen.");

  addItem("Free Cash Flow", 2.0, s1.freeCashFlow, s2.freeCashFlow, true, fmtLargeNum,
    "Actual cash left after capital expenditures. The lifeblood of the business — funds buybacks, dividends, acquisitions, and R&D without diluting shareholders.");

  // ── Tier 4: Financial health ─────────────────────────────────────────────
  addItem("Gross Margin", 1.5, s1.grossMargin, s2.grossMargin, true, fmtPct,
    "Revenue minus direct costs. Software/platform companies typically run >60%; below 30% signals a commoditised business with little pricing power.");

  addItem("Debt-to-Equity", 1.5, s1.debtToEquity, s2.debtToEquity, false, fmtNum,
    "Total debt divided by shareholder equity. D/E < 0.5 is conservatively financed; > 2.0 warrants scrutiny. High leverage amplifies both gains and losses.");

  addItem("Current Ratio", 1.0, s1.currentRatio, s2.currentRatio, true, fmtNum,
    "Current assets ÷ current liabilities. > 1.5 is healthy; < 1.0 means the company cannot cover near-term obligations from existing assets alone.");

  const ticker1TotalScore = items.reduce((acc, it) => acc + it.ticker1Score * it.weight, 0);
  const ticker2TotalScore = items.reduce((acc, it) => acc + it.ticker2Score * it.weight, 0);
  const maxScore = items.reduce((acc, it) => acc + it.weight, 0);

  const winnerTicker = ticker1TotalScore >= ticker2TotalScore ? ticker1 : ticker2;
  const loserTicker = winnerTicker === ticker1 ? ticker2 : ticker1;
  const winnerName = winnerTicker === ticker1 ? s1.companyName : s2.companyName;
  const gap = Math.abs(ticker1TotalScore - ticker2TotalScore) / maxScore;
  const confidence = gap > 0.25 ? "High" : gap > 0.1 ? "Medium" : "Low";

  const winnerScore = Math.max(ticker1TotalScore, ticker2TotalScore);
  const loserScore = Math.min(ticker1TotalScore, ticker2TotalScore);

  const summary = `${winnerName} (${winnerTicker}) scores ${winnerScore.toFixed(1)} vs. ${loserScore.toFixed(1)} for ${loserTicker}, making it the stronger pick with ${confidence.toLowerCase()} conviction. ` +
    `${winnerTicker} shows advantages in ` +
    items.filter(i => i.winner === winnerTicker).slice(0, 3).map(i => i.metric).join(", ") + ".";

  return {
    ticker1TotalScore: parseFloat(ticker1TotalScore.toFixed(2)),
    ticker2TotalScore: parseFloat(ticker2TotalScore.toFixed(2)),
    winner: winnerName,
    winnerTicker,
    loserTicker,
    confidence,
    summary,
    items,
  };
}

function buildMetrics(quote: any, ticker: string) {
  // Yahoo Finance v3: trailingEps / forwardEps (not epsTrailingTwelveMonths / epsForward)
  const eps = safeNum(quote.trailingEps);
  const epsForward = safeNum(quote.forwardEps);
  const epsGrowth = eps !== null && epsForward !== null && eps > 0
    ? (epsForward - eps) / Math.abs(eps)
    : null;

  // Prefer trailing P/E; fall back to forward P/E only if trailing is absent
  const peRatio = safeNum(quote.trailingPE) ?? safeNum(quote.forwardPE);

  // Yahoo Finance returns revenueGrowth / earningsGrowth as decimals (0.84 = 84%)
  const revenueGrowthYoY = safeNum(quote.revenueGrowth);
  // earningsGrowth (TTM) is Yahoo's best proxy for near-term growth momentum
  const revenueGrowthProjected = safeNum(quote.earningsGrowth);

  const dividendYield = safeNum(quote.dividendYield);

  // Prefer Yahoo's own PEG; fall back to computed P/E ÷ annualised EPS growth %
  const pegRatio = safeNum(quote.pegRatio) ??
    (peRatio !== null && epsGrowth !== null && epsGrowth > 0
      ? peRatio / (epsGrowth * 100)
      : null);

  const stockType = classifyStockType(peRatio, pegRatio, dividendYield, revenueGrowthYoY);
  const fairValue = computeFairValue(eps, epsGrowth, peRatio);

  // Yahoo Finance debtToEquity comes as a percentage figure (e.g. 247.7 means D/E ratio 2.477)
  const debtToEquityRaw = safeNum(quote.debtToEquity);
  const debtToEquity = debtToEquityRaw !== null ? debtToEquityRaw / 100 : null;
  const leverageRatio = debtToEquity !== null ? debtToEquity + 1 : null;

  return {
    ticker: ticker.toUpperCase(),
    companyName: quote.longName ?? quote.shortName ?? ticker.toUpperCase(),
    currentPrice: safeNum(quote.regularMarketPrice),
    marketCap: safeNum(quote.marketCap),
    peRatio,
    pegRatio,
    priceToBook: safeNum(quote.priceToBook),
    priceToSales: safeNum(quote.priceToSalesTrailing12Months),
    leverageRatio,
    debtToEquity,
    totalRevenue: safeNum(quote.totalRevenue),
    revenueGrowthYoY,
    revenueGrowthProjected,
    netIncome: safeNum(quote.netIncomeToCommon),
    ebitda: safeNum(quote.ebitda),
    earningsPerShare: eps,
    epsGrowth,
    freeCashFlow: safeNum(quote.freeCashflow),
    dividendYield,
    returnOnEquity: safeNum(quote.returnOnEquity),
    returnOnAssets: safeNum(quote.returnOnAssets),
    currentRatio: safeNum(quote.currentRatio),
    grossMargin: safeNum(quote.grossMargins),
    operatingMargin: safeNum(quote.operatingMargins),
    netMargin: safeNum(quote.profitMargins),
    beta: safeNum(quote.beta),
    fiftyTwoWeekHigh: safeNum(quote.fiftyTwoWeekHigh),
    fiftyTwoWeekLow: safeNum(quote.fiftyTwoWeekLow),
    analystTargetPrice: safeNum(quote.targetMeanPrice),
    fairValueEstimate: fairValue,
    stockType,
    sector: quote.sector ?? null,
    industry: quote.industry ?? null,
    exchange: quote.exchangeName ?? quote.exchange ?? null,
    currency: quote.currency ?? "USD",
    logoUrl: null,
    description: quote.longBusinessSummary?.slice(0, 300) ?? null,
  };
}

router.get("/stocks/compare", async (req, res) => {
  const parsed = CompareStocksQueryParams.safeParse(req.query);
  if (!parsed.success) {
    return res.status(400).json({ error: "ticker1 and ticker2 are required" });
  }
  const { ticker1, ticker2 } = parsed.data;
  const cacheKey = `${ticker1.toUpperCase()}:${ticker2.toUpperCase()}`;

  const cached = compareCache.get(cacheKey);
  if (cached) return res.json(cached);

  try {
    const [q1, q2] = await Promise.all([
      yahooFinance.quoteSummary(ticker1.toUpperCase(), {
        modules: [
          "price",
          "summaryDetail",
          "financialData",
          "defaultKeyStatistics",
          "assetProfile",
        ],
      }),
      yahooFinance.quoteSummary(ticker2.toUpperCase(), {
        modules: [
          "price",
          "summaryDetail",
          "financialData",
          "defaultKeyStatistics",
          "assetProfile",
        ],
      }),
    ]);

    const merged1 = {
      ...q1.price,
      ...q1.summaryDetail,
      ...q1.financialData,
      ...q1.defaultKeyStatistics,
      ...q1.assetProfile,
    };
    const merged2 = {
      ...q2.price,
      ...q2.summaryDetail,
      ...q2.financialData,
      ...q2.defaultKeyStatistics,
      ...q2.assetProfile,
    };

    const stock1 = buildMetrics(merged1, ticker1);
    const stock2 = buildMetrics(merged2, ticker2);
    const scorecard = buildScorecard(stock1, stock2, ticker1.toUpperCase(), ticker2.toUpperCase());

    const payload = { stock1, stock2, scorecard };
    compareCache.set(cacheKey, payload);
    return res.json(payload);
  } catch (err: any) {
    const msg = String(err?.message ?? err);
    if (msg.includes("No fundamentals data") || msg.includes("Not Found")) {
      return res.status(404).json({ error: `Ticker not found: ${msg}` });
    }
    return res.status(500).json({ error: `Failed to fetch stock data: ${msg}` });
  }
});

router.get("/stocks/search", async (req, res) => {
  const parsed = SearchStocksQueryParams.safeParse(req.query);
  if (!parsed.success) {
    return res.status(400).json({ error: "q is required" });
  }
  const { q } = parsed.data;
  const cacheKey = q.toUpperCase();

  const cached = searchCache.get(cacheKey);
  if (cached) return res.json(cached);

  try {
    const results = await yahooFinance.search(q, { newsCount: 0, quotesCount: 8 });
    const quotes = (results.quotes ?? [])
      .filter((r: any) => r.isYahooFinance && r.symbol)
      .slice(0, 8)
      .map((r: any) => ({
        ticker: r.symbol,
        name: r.longname ?? r.shortname ?? r.symbol,
        exchange: r.exchange ?? "",
        type: r.quoteType ?? "EQUITY",
      }));
    searchCache.set(cacheKey, quotes);
    return res.json(quotes);
  } catch (err: any) {
    return res.status(500).json({ error: String(err?.message ?? err) });
  }
});

const quoteCache = new TTLCache<object>(10 * 60 * 1000);

async function fetchQuoteSummary(ticker: string) {
  return yahooFinance.quoteSummary(ticker.toUpperCase(), {
    modules: ["price", "summaryDetail", "financialData", "defaultKeyStatistics", "assetProfile"],
  });
}

router.get("/stocks/quote", async (req, res) => {
  const { ticker } = req.query as { ticker?: string };
  if (!ticker || typeof ticker !== "string") {
    return res.status(400).json({ error: "ticker is required" });
  }
  const key = ticker.toUpperCase();
  const cached = quoteCache.get(key);
  if (cached) return res.json(cached);

  try {
    const q = await fetchQuoteSummary(key);
    const merged = { ...q.price, ...q.summaryDetail, ...q.financialData, ...q.defaultKeyStatistics, ...q.assetProfile };
    const metrics = buildMetrics(merged, key);
    quoteCache.set(key, metrics);
    return res.json(metrics);
  } catch (err: any) {
    const msg = String(err?.message ?? err);
    if (msg.includes("No fundamentals data") || msg.includes("Not Found")) {
      return res.status(404).json({ error: `Ticker not found: ${key}` });
    }
    return res.status(500).json({ error: `Failed to fetch stock data: ${msg}` });
  }
});

export default router;