# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stock Comparison Tool API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface StockSearchResult {
  ticker: string;
  name: string;
  exchange?: string;
  type?: string;
}

export interface StockMetrics {
  ticker: string;
  companyName: string;
  /** @nullable */
  currentPrice?: number | null;
  /** @nullable */
  marketCap?: number | null;
  /** @nullable */
  peRatio?: number | null;
  /** @nullable */
  pegRatio?: number | null;
  /** @nullable */
  priceToBook?: number | null;
  /** @nullable */
  priceToSales?: number | null;
  /** @nullable */
  leverageRatio?: number | null;
  /** @nullable */
  debtToEquity?: number | null;
  /** @nullable */
  totalRevenue?: number | null;
  /** @nullable */
  revenueGrowthYoY?: number | null;
  /** @nullable */
  revenueGrowthProjected?: number | null;
  /** @nullable */
  netIncome?: number | null;
  /** @nullable */
  ebitda?: number | null;
  /** @nullable */
  earningsPerShare?: number | null;
  /** @nullable */
  epsGrowth?: number | null;
  /** @nullable */
  freeCashFlow?: number | null;
  /** @nullable */
  dividendYield?: number | null;
  /** @nullable */
  returnOnEquity?: number | null;
  /** @nullable */
  returnOnAssets?: number | null;
  /** @nullable */
  currentRatio?: number | null;
  /** @nullable */
  grossMargin?: number | null;
  /** @nullable */
  operatingMargin?: number | null;
  /** @nullable */
  netMargin?: number | null;
  /** @nullable */
  beta?: number | null;
  /** @nullable */
  fiftyTwoWeekHigh?: number | null;
  /** @nullable */
  fiftyTwoWeekLow?: number | null;
  /** @nullable */
  analystTargetPrice?: number | null;
  /** @nullable */
  fairValueEstimate?: number | null;
  /**
     * Classified type - Growth, Value, Dividend, Blend
     * @nullable
     */
  stockType?: string | null;
  /** @nullable */
  sector?: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  exchange?: string | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface ScorecardItem {
  metric: string;
  weight: number;
  ticker1Score: number;
  ticker2Score: number;
  /** @nullable */
  ticker1Value?: string | null;
  /** @nullable */
  ticker2Value?: string | null;
  /** @nullable */
  winner?: string | null;
  explanation: string;
}

export interface Scorecard {
  ticker1TotalScore: number;
  ticker2TotalScore: number;
  winner: string;
  winnerTicker: string;
  loserTicker: string;
  /** High, Medium, or Low based on score gap */
  confidence: string;
  summary: string;
  items: ScorecardItem[];
}

export interface StockComparison {
  stock1: StockMetrics;
  stock2: StockMetrics;
  scorecard: Scorecard;
}

export type GetStockQuoteParams = {
/**
 * Stock ticker symbol (e.g. AAPL)
 */
ticker: string;
};

export type CompareStocksParams = {
/**
 * First stock ticker symbol (e.g. AAPL)
 */
ticker1: string;
/**
 * Second stock ticker symbol (e.g. MSFT)
 */
ticker2: string;
};

export type SearchStocksParams = {
/**
 * Search query (company name or ticker)
 */
q: string;
};